refactoring - statistics view was separated
